fix : Moved extra code to new file
